requirements-dev.txt
# 开发环境额外依赖
-r requirements.txt

# 开发工具
ipython==8.17.2
jupyter==1.0.0
jupyter-lab==4.0.9

# 代码质量工具
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
autoflake==2.2.1
yapf==0.40.2

# 测试工具
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-html==4.1.1
pytest-benchmark==4.0.0
factory-boy==3.3.0
faker==20.1.0
responses==0.24.1

# 调试工具
pdb++==0.10.3
rich==13.7.0
icecream==2.1.3

# 性能分析
py-spy==0.3.14
memory-profiler==0.61.0
line-profiler==4.1.1
snakeviz==2.2.0

# 文档生成
mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocs-mermaid2-plugin==1.1.1
mkdocstrings==0.24.0

# 代码分析
radon==6.0.1
xenon==0.9.1
bandit==1.7.5
safety==2.3.5
licensecheck==2023.3
pydocstyle==6.3.0

# 类型检查
mypy==1.7.1
types-redis==4.6.0.11
types-requests==2.31.0.10
types-PyYAML==6.0.12.12
types-setuptools==69.0.0.0

# 热重载
watchdog==3.0.0
watchfiles==0.21.0
watchgod==0.8.2

# 环境管理
python-dotenv==1.0.0
envparse==0.2.0

# 数据库工具
sqlalchemy-stubs==0.4
alembic-utils==0.8.2

# API测试
httpx==0.25.2
requests==2.31.0
requests-mock==1.11.0

# 模拟工具
responses==0.24.1
freezegun==1.2.2

# 覆盖率报告
coverage==7.3.2
coverage-badge==1.1.0

# 依赖管理
pipdeptree==2.13.1
pip-tools==7.3.0

# 代码格式化配置
toml==0.10.2

# 开发服务器
uvicorn[standard]==0.24.0

# 数据库迁移
alembic==1.12.1

# 配置文件验证
cerberus==1.3.5
voluptuous==0.14.1

# 日志美化
colorama==0.4.6
termcolor==2.4.0

# 进度条
tqdm==4.66.1

# 时间处理
arrow==1.3.0
pendulum==2.1.2

# 数据生成
mimesis==11.1.0

# 网络工具
networkx==3.2.1

# 文件处理
pathlibutil==0.2.12

# 异步工具
aiofiles==23.2.1
aiohttp==3.9.1
aioredis==2.0.1

# 缓存工具
cachetools==5.3.2

# 序列化
orjson==3.9.10
msgpack==1.0.7

# 压缩
zstandard==0.22.0
lz4==4.3.2

# 图像处理
Pillow==10.1.0

# 数据处理
pandas==2.1.4
numpy==1.25.2

# 可视化
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# 机器学习
scikit-learn==1.3.2

# 文本处理
jieba==0.42.1

# 加密
cryptography==41.0.7

# 验证
email-validator==2.1.0

# 配置
pydantic==2.5.0
pydantic-settings==2.1.0

# 命令行
click==8.1.7
typer==0.9.0

# 系统信息
psutil==5.9.6

# 容器
docker==6.1.3

# 云服务
boto3==1.34.0

# 监控
prometheus-client==0.19.0

# 错误追踪
sentry-sdk[fastapi]==1.38.0

# 性能监控
newrelic==9.2.0

# 健康检查
py-healthcheck==1.10.1

# 开发辅助
ipdb==0.13.13
ptpython==3.0.25
bpython==0.24

# 代码搜索
grep==0.3.1

# 文件同步
watchdog==3.0.0

# 数据库GUI
sqlitebrowser==0.12.2

# API文档
redoc==2.1.3
swagger-ui-bundle==0.0.9

# 开发配置
setuptools==69.0.2
wheel==0.42.0
twine==4.0.2

# 版本管理
bump2version==1.0.1
semver==3.0.2

# 许可证
license==0.1.0

# 代码统计
cloc==0.2.5

# 依赖图
pipdeptree==2.13.1

# 安全审计
pip-audit==2.6.2

# 代码重构
rope==1.11.0

# 导入优化
unusedcode==1.1.0

# 代码复杂度
mccabe==0.7.0

# 文档字符串
interrogate==1.5.0

# 类型存根
data-science-types==0.2.23

# 异步测试
pytest-asyncio==0.21.1
aioresponses==0.7.6

# 并发测试
pytest-xdist==3.5.0

# 基准测试
pytest-benchmark==4.0.0

# 内存测试
pytest-memray==1.0.0

# 超时测试
pytest-timeout==2.2.0

# 随机测试
pytest-randomly==3.15.0

# 重试测试
pytest-rerunfailures==12.0

# 并行测试
pytest-parallel==0.1.1

# 测试报告
pytest-html==4.1.1
pytest-json-report==1.5.0

# 测试覆盖率
pytest-cov==4.1.0

# 测试数据
pytest-factoryboy==2.5.1

# 测试模拟
pytest-mock==3.12.0

# 测试快照
syrupy==4.6.1

# 测试固件
pytest-fixtures==0.1.0

# 测试参数化
pytest-cases==3.8.2

# 测试标记
pytest-marker==1.0.0

# 测试插件
pytest-plugins==1.0.0

# 测试工具
pytest-toolbox==0.4

# 测试助手
pytest-helpers==0.2.0

# 测试实用工具
pytest-utils==1.0.0